requirements.txt
# Core Python Packages
pyspark==3.5.1           # or latest stable version
boto3==1.34.78
pymongo[srv]==4.7.2
pandas==2.2.2
python-dotenv==1.0.1
